backend/requirements.txt
# FastAPI and web server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image processing dependencies
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Machine learning dependencies
torch==2.1.1
torchvision==0.16.1
ultralytics==8.3.176
joblib==1.3.2
numpy==1.24.4
pandas==2.1.3
scikit-learn==1.3.2

# Additional utilities
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.28.1
pytorch-tabnet==4.1.0
requests==2.31.0
